수정중...

+++ src/main/java/com/lyj/securitydomo/dto/ReportDTO.java
@@ -12,9 +12,9 @@ import java.util.Date;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ReportDTO {
-    private Integer reportId;
-    private Integer postId; // Post와의 관계
-    private Integer userId; // User와의 관계
+    private Long reportId;
+    private Long postId; // Post와의 관계
+    private Long userId; // User와의 관계
     private String reason; // 신고 사유
     private String status; // 신고 진행 상태
     private String category; // 신고 분류
